backend/testing_scripts/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
python-jose[cryptography]
passlib[bcrypt]
requests